Fixed rollback bug in GitPreparer
Added a guard to GitPreparer::unsetVcsInProgress() to prevent doing a
`git reset` when there are no test locks present.

Otherwise, the uncommitted changes in the app will be removed by the
shutdown feature introduced in 952c6e5890daa36236be675d8c4f21cecabc1fe7.

+++ lib/preparers/GitPreparer.php
@@ -7,11 +7,13 @@ class GitPreparer implements Preparer
 
 	public function snapshot()
 	{
+		$this->debug('Snapshot requested');
 		return $this->setVcsInProgress();
 	}
 
 	public function rollback()
 	{
+		$this->debug('Rollback requested');
 		return $this->unsetVcsInProgress();
 	}
 
@@ -88,6 +90,12 @@ class GitPreparer implements Preparer
 
 	protected function unsetVcsInProgress()
 	{
+		if (!$this->isVcsInProgress())
+		{
+			$this->debug('No test locks found');
+			return;
+		}
+
 		$this->debug('Rolling back Git repo to pre-test state…');
 		$this->runCommand('git reset --hard HEAD');
 		$this->runCommand('git clean -fdx');
